requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
certifi==2022.6.15
charset-normalizer==2.1.0
flake8==5.0.2
frozenlist==1.3.0
idna==3.3
jarowinkler==1.2.0
mccabe==0.7.0
multidict==6.0.2
pycodestyle==2.9.0
pyflakes==2.5.0
pytz==2022.1
yarl==1.7.2
python-dotenv==0.20.0
plotly
kaleido
gspread
pandas